update: update user.service
